fix display number books of each type
