+++ src/main/webapp/WEB-INF/views/elements/admin/category.jsp
@@ -10,7 +10,7 @@
 	<table class="category-list table">
 		<div class="filter-wrapper row">
 			<div class="col-md-10 col-sm-12 col-lg-10">
-				<form class="d-flex" role="search" action="/admin/category/search.htm">
+				<form role="search" action="admin/category/search.htm">
 					<input name="searchInput" class="form-control" type="search" placeholder="Tìm thể loại" aria-label="Search" style="width: 50%;">
 				</form>
 			</div>
@@ -30,7 +30,7 @@
 			<c:forEach var="category" items="${categories}">
 				<tr>
 					<td>${category.TENTL}</td>
-					<td>3</td>
+					<td>${category.getBooks().size()}</td>
 					<td>100.000</td>
 					<td>
 						<a href="admin/category/${category.MATL}/delete.htm" onclick="return confirm('Bạn có chắc muốn xóa thể loại ${category.TENTL}?')"><button class="btn btn-outline-primary"><i class="bi bi-trash3-fill"></i></button></a>
